requirements.txt
apscheduler==3.10.4
beautifulsoup4==4.12.3
httpx==0.27.2
lxml==5.3.0
orjson==3.10.7
pydantic==2.8.2
pymongo[srv]==4.8.0
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
scrapy==2.11.2
structlog==24.1.0
tenacity==9.0.0
rapidfuzz==3.9.7
pytest==8.3.3
ruff==0.6.9
playwright==1.48.0
playwright==1.48.0
